+++ app/src/main/java/com/prakash/pwall/service/PWallWallpaperService.kt
@@ -95,6 +95,12 @@ class PWallWallpaperService : WallpaperService() {
                     renderOnce()
                 }
             }
+            scope.launch {
+                container.maskEditNotifier.collect {
+                    depthEngine.reloadFromCache()
+                    renderOnce()
+                }
+            }
         }
 
         override fun onDestroy() {